% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{as.list.asg}
\alias{as.list.asg}
\title{`as.list.asg` return a list representation for an asg graph object}
\usage{
\method{as.list}{asg}(x, ...)
}
\arguments{
\item{x}{an asg graph object created with asg.new}

\item{\ldots}{additional arguments, delegated to the list command}
}
\value{
list object with the components theta, sigma, p.value
}
\description{
`as.list.asg` provides a S3 method to convert an asg graph object
             into a list object which can be for instance used to write a report using the library openxlsx.
}
\examples{
data(swiss)
as=asg.new(swiss,method="spearman",alpha=0.1)
result=as.list(as)
ls(result)
result$settings
# can be writte as xlsx file for instance like:
# library(openxlsx)
# write.xlsx(result,file="some-result.xlsx")
}
\seealso{
\link[asg:plot.asg]{plot.asg}
}
