% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.bench}
\alias{asg.bench}
\title{Measure time and quality of an graph prediction.}
\usage{
asg.bench(g, FUN, ...)
}
\arguments{
\item{g}{the true graph, either an asg graph object or and adjacency matrix}

\item{FUN}{the function to do the prediction, should only return the adjacency matrix}

\item{\ldots}{additional arguments passed to FUN}
}
\value{
list with the components time, SENS, SPEC and MCC
}
\description{
This is a utility function to measure the executation time
and the quality of the graph prediction using the sensitivity, specificity and
Matthews Correlation coefficient.
}
\examples{
set.seed(123)
W=mgraph.new(type="werner")
data=t(mgraph.graph2data(W,iter=50,prop=0.02))
res = asg.bench(W,FUN=function() { asg.new(data,method="spearman",alpha=0.1) })
unlist(res)
res = asg.bench(W,FUN=function() { asg.new(data,method="spearman",alpha=0.1,prob=TRUE) })
unlist(res)
AN=mgraph.new(type="angie",nodes=20,edges=30)
plot(AN)
data=t(mgraph.graph2data(AN))
unlist(asg.bench(AN, function () { 
   return(asg.new(data,method="spearman",alpha=0.1,prob=FALSE)$theta) }))
unlist(asg.bench(AN, function () { 
   return(asg.new(data,method="spearman",alpha=0.1,prob=TRUE)$theta) }))
}
