% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.cm}
\alias{asg.cm}
\title{Check graph for edge additions using linear models.}
\usage{
asg.cm(graph, rs.threshold = 0.02, .iter = 0)
}
\arguments{
\item{graph}{an asg graph object}

\item{rs.threshold}{threshold for the required increase in the adjusted r-square value}

\item{.iter}{internal variable to measure the number of iterations, 
after 5 iterations the process is stopped}
}
\value{
an asg graph object with possible added edges
}
\description{
This is a utility function to add edges if the linear
model for a node, which is created by using connected nodes, can be
improved by adding other nodes which have significant correlations pairwise
correlations.
}
\examples{
set.seed(123)
par(mfrow=c(2,2),mai=rep(0.4,4))
W=mgraph.new(type="werner")
plot(W,layout="circle")
data=mgraph.graph2data(W,iter=100,prop=0.02)
A = asg.new(t(data),method="spearman",alpha=0.1)
plot(A,layout="circle")
B = asg.cm(A)
plot(B,layout="circle")
}
