% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.getChains}
\alias{asg.getChains}
\title{Return the chains of an asg graph as data frame}
\usage{
asg.getChains(graph)
}
\arguments{
\item{graph}{an asg graph object}
}
\value{
A matrix with one chain per row. Shorter chains are filled up with empty strings.
}
\description{
This is a utiltiy function to return the chains which constructs the graph as a matrix.
}
\examples{
data(swiss)
sw.g=asg.new(swiss)
asg.getChains(sw.g)
}
