% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.getFactors}
\alias{asg.getFactors}
\title{Performs a simple factor analysis and performs the Factor ids.}
\usage{
asg.getFactors(x, factors = 0, na.omit = TRUE, ...)
}
\arguments{
\item{x}{A formula or a numeric matrix or an object that can be coerced to a numeric matrix.}

\item{factors}{The number of factor to be fitted. If 0 is given the function will select the number of factors which explain more than 90\% of the variablity of the data set or the the maximum possible number of factors.}

\item{na.omit}{What do with missing values, if TRUE all rows and columns with missing values will be deleted, if FALSE an error is thrown, default: FALSE}

\item{\dots}{Arguments delegated to the factanal function.}
}
\value{
A list object with factor ids, proportional variance and the factanal object.
}
\description{
This function returns a list with two elements the factor analysis ids for each variable and a factor analysis object of class factanal.
}
\examples{
data(swiss)
fact=asg.getFactors(swiss)
fact$ids
fact$var
plot(asg.new(swiss),hilight.factors=TRUE,factors=2)
}
