% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.getGraphAccuracy}
\alias{asg.getGraphAccuracy}
\title{Measure similarity between two graphs based on edges}
\usage{
asg.getGraphAccuracy(true.graph, pred.graph)
}
\arguments{
\item{true.graph}{an qadjacency matrix or an asg graph object}

\item{pred.graph}{an adjacency matrix or an asg graph object}
}
\value{
A list object with measures such as accuracy, sensitivity, specificity, balanced classification rate and F-score.
}
\description{
This function returns similarity measures between two graphs based on predictions of edges. 
The accuracy is the proportion of correctly predicted edges from all predicted edges. Also other measures, sensitivity, specificity, balanced classification rates and F1 score are releated to the predicted edges.
}
\examples{
data(swiss)
sw.g=asg.new(swiss)
G=sw.g$theta # assume that this is the true graph
G
H=sw.g$sigma^2 # create a simple threshold based graph
H[H>0.1]=1
H[H<=0.1]=0
diag(H)=0
H
asg.getGraphAccuracy(G,H)
}
