% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.gui}
\alias{asg.gui}
\title{Start a graphical user interface for the SNH algorithm}
\usage{
asg.gui(wd = file.path(system.file(package = "asg"), "exec"))
}
\arguments{
\item{wd}{a matrix or a dataframe with variables in columns, default: exec directory of the package}
}
\description{
This function is a simple wrapper function to download the actual version of the GUI
   for performing an analysis with the St- Nicolas House Algorithm. After execution a GUI window should popup. Please read the help pages
   in the Help tab to find out how to use this application.
}
\examples{
\dontrun{
  asg.gui()
}
}
