% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R, R/impute.R
\name{asg.impute}
\alias{asg.impute}
\title{Imputation of missing values using rpart, knn, mean or median}
\usage{
asg.impute(x, method = "knn", k = 5, cor.method = "spearman")

asg.impute(x, method = "knn", k = 5, cor.method = "spearman")
}
\arguments{
\item{x}{either a matrix or data frame}

\item{method}{method used for replacing missing values, either mean, median, rpart or knn, default 'rpart'}

\item{k}{number of neighbors for imputation if using 'knn', default: 5}

\item{cor.method}{correlation method for calcuting distance if using 'knn', default: 'spearman'}
}
\value{
data with imputed values.

depending on the input either a data frame or matrix with NA's replaced by imputed values.
}
\description{
`asg.impute` imputes missing data in data frames or matrices using either
  decision trees, k-nearest-neighbor approach with Euclidean distances or
  simple mean and median computations for the variables.

Replaces missing values with a reasonable guess by different imputation methods such as 
  the simple and not recommended methods mean and median, where NA's are replaced with the 
  mean or median for this variable or the more recommended methods using rpart decision trees
  or knn using a correlation distance based k-nearest neighbor approach. 
  The rpart method can be as well used to replace missing values for categorical variables.
  In case of median and mean imputations for categorical variables the modus is used, 
  so missing values are replaced with the most often category. This is rarely reasonable.
}
\section{Details}{

    Many mathematical methods did not allow missing data in their inputs.
    The missing values have to be guessed in this case.
    The more basic approaches are replacing the NA values with the median or mean for the variable.
    More advanced methods are using the mean only for a few samples which are very similar to the sample where the values is missing.
    This method is called knn-imputation and uses the k-nearest neighbors (default here 5) only for computing the replacement value.
    For data where the value order in the columns is of importance it is often desired to replace missing values with the mean of the two closest neighbors in the data frame or matrix, for instance if the data are ordered by time. Here the method timemean follows this approach.
    The method rpart uses decision trees to impute the values, it is currently the only method to impute as well factor variables.  The advantage of the rpart method is that it can impute not only numerical values but as well factor variables.
}

\examples{
data(iris)
set.seed(123)
ir=as.matrix(iris[,1:4])
ir.mv=ir
# introduce 5 percent NA's
mv=sample(1:length(ir),as.integer(0.05*length(ir)))
ir.mv[mv]=NA
ir.imp.med=asg.impute(ir.mv,method='median') # not good
ir.imp.rpart=asg.impute(ir.mv,method="knn") # method rpart 
ir.imp.knn=asg.impute(ir.mv,method='knn') # (default)
rmse = function (x,y) { return(sqrt(sum((x-y)^2))) }
rmse(ir[mv],ir.imp.med[mv]) # should be high
rmse(ir[mv],ir.imp.rpart[mv]) # should be low!
rmse(ir[mv],ir.imp.knn[mv]) # should be low!
cor(ir[mv],ir.imp.med[mv])
cor(ir[mv],ir.imp.rpart[mv])
cor(ir[mv],ir.imp.knn[mv]) # should be high!
# factor variables
data(iris)
ciris=iris
idx=sample(1:nrow(ciris),15) # 10 percent NA's
ciris$Species[idx]=NA
summary(ciris)
ciris=asg.impute(ciris,method="rpart")
table(ciris$Species[idx],iris$Species[idx])
data(iris)
ir=as.matrix(iris[,1:4])
irc=ir
idx=sample(1:length(irc),50)
irc[idx]=NA
summary(irc)
irp=asg.impute(irc,method="rpart")
irk=asg.impute(irc,method="knn")
irm=asg.impute(irc,method="mean")
idx=which(is.na(irc))
M=data.frame(orig=ir[idx],rpart=irp[idx],knn=irk[idx],mean=irm[idx])
round(cor(M),2)

}
\seealso{
\link{asg.new}.
}
\author{
Detlef Groth <email: dgroth@uni-potsdam.de>
}
