% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.layout}
\alias{asg.layout}
\title{Determine graph layouts.}
\usage{
asg.layout(
  A,
  mode = "sam",
  method = "pearson",
  noise = FALSE,
  star.center = NULL,
  interactive = FALSE
)
}
\arguments{
\item{A}{An adjacency matrix or an asg graph object.}

\item{mode}{Character string for the layout type, can be either 'mds' (mds on graph), 'mdsd' (mds on data) 'sam' (sammon on graph), 'samd' (sammon on data), 'circle', 'grid' or 'star', default: 'sam'}

\item{method}{method for calculating correlation distance if mode is either 'mdsd' or 'samd', default: pearson}

\item{noise}{Should some noise be added, default: FALSE.}

\item{star.center}{the centered node if layout is 'star', must be a character string for the nodename, default: NULL.}

\item{interactive}{switch into interactive mode where you can click in the graph and move nodes with two clicks, first selecting the node, second click gives thehe new coordinates for the node, default: FALSE}
}
\value{
A matrix with x and y columns for the layout.
}
\description{
This function returns xy coordinates for a given input adjacency matrix or asg graph. 
This function is useful if you like to plot the same set of nodes with different edge connections 
in the same layout.
}
\examples{
data(swiss)
sw.s=asg.new(swiss,method='spearman')
sw.p=asg.new(swiss,method='pearson')
lay=asg.layout(sw.s,mode='sam')
plot(sw.s,layout=lay)
plot(sw.p,layout=lay)
plot(sw.s,layout='star',star.center='Education')
rn1=rnorm(nrow(swiss))
nswiss=cbind(swiss,Rn1=rn1)
plot(asg.new(nswiss,method='spearman'),layout='sam')
plot(asg.new(nswiss,method='spearman'),layout='samd',
  vertex.size=2,vertex.color='beige')
}
\keyword{layout}
\keyword{network}
