% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.ll}
\alias{asg.ll}
\title{log-likelihood for the given asg graph and the given chain}
\usage{
asg.ll(g, chain = NULL)
}
\arguments{
\item{g}{an asg graph object}

\item{chain}{a chain object of an asg graph, if not given a data frame with the values is returned for all chains, default: NULL}
}
\value{
list with the following components: ll.total, ll.chain, ll.rest, ll.block
      df, chisq, p.value,  block.df, block.ch, block.p.value. If chain is not given an overal summary is made for all chains an returned as data frame.
}
\description{
This function returns the log-likelihood for the given asg graph and the given chain.
}
\examples{
data(swiss)
sw.g=asg.new(swiss)
asg.ll(sw.g,sw.g$chain$Catholic)
head(asg.ll(sw.g))
}
