% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.mdsplot}
\alias{asg.mdsplot}
\title{MDS plot for an asg object}
\usage{
asg.mdsplot(
  x,
  mds.method = cmdscale,
  dist.method = function(x) {
     1 - abs(x)
 },
  grid = TRUE,
  pch = 19,
  col = "salmon",
  ...
)
}
\arguments{
\item{x}{object of class asg, created using the function asg.new}

\item{mds.method}{the method to compute the new coordinate system, default: cmdscale}

\item{dist.method}{the method to compute the distance, default: 1-abs(sigma)}

\item{grid}{should a grid plotted, default: TRUE}

\item{pch}{plotting character, default: 19}

\item{col}{plotting color, default: salmon}

\item{\ldots}{additional arguments delegated to the standard plot function}
}
\description{
The function `asg.mdsplot` provides an MDS plot for
  a object created with the asg.new command. It uses the correlation distance as
  d(a,b)=1-abs(r(a,b)), so the same data which are used to generate the associatian
  chaing graph.
}
\examples{
library(MASS)
data(birthwt)
par(mfrow=c(1,2))
as=asg.new(birthwt[,2:ncol(birthwt)],method="spearman")
plot(as,layout="sam")
asg.mdsplot(as)   # high negative correlations are close to each other per default
asg.mdsplot(as,dist.method=function(x) { (1-x)/2 }) # negative correlations are far
}
\seealso{
\link[asg:plot.asg]{plot.asg}
}
