% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.mplot}
\alias{asg.mplot}
\title{`asg.mplot` using the asg plot for simple adjacency matrices}
\usage{
asg.mplot(M, main = "", ...)
}
\arguments{
\item{M}{an adjacency matrix, or an asg graph object}

\item{main}{title of the plot, Default: empty string}

\item{\dots}{all other parameters delegated to plot.asg}
}
\value{
nothing.
}
\description{
`asg.mplot` provides the plot.asg functions as well for simple adjacency matrices
}
\section{Details}{

This is a small plot utility to display networks or correlation matrices of asg graph objects.
}

\examples{
M=matrix(rbinom(100,1, 0.2),nrow=10,ncol=10)
diag(M)=0
colnames(M)=rownames(M)=LETTERS[1:10]
asg.mplot(M)
}
\seealso{
\link[asg:plot.asg]{plot.asg}
}
