% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.nd}
\alias{asg.nd}
\title{network deconvolution based on Feizi et. al. 2013}
\usage{
asg.nd(x, beta = 0.99, alpha = 1, control = FALSE)
}
\arguments{
\item{x}{symmetric relevance matrix, where high similarity between nodes is expressed with high values such as in a correlation matrix}

\item{beta}{scaling paramater, the largest absolute eigenvalue is mapped to beta, values should be between 0 and 1, default: 0.99}

\item{alpha}{fraction of edges of the observed dependency matrix to be kept, should be between 0 (none) and 1 (all), default: 1}

\item{control}{if FALSE only direct weights are displayed, if TRUE also non-observed interactions are displayed, default: FALSE}
}
\description{
Return a network deconvolved data matrix using the algorithm of Feizi et. a. 2013.
This function is a R implementation for "A General Method to Distinguish Direct Dependencies over Networks" by Feizi et. al (2013). For the Matlab code look here:
\url{http://compbio.mit.edu/nd/code/ND.m}
}
\details{
<a name="nd"> </a>
}
\examples{
options(width=100)
data(swiss)
swissr=cbind(swiss,Rand=rnorm(nrow(swiss)))
mat=cor(swissr,method="pearson")
round(mat,2)
round(abs(mat),2)
round(asg.nd(abs(mat)),2)
round(asg.nd(abs(mat),alpha=0.2,beta=0.9),2)
round(asg.nd(abs(mat),alpha=0.2,beta=0.9,control=TRUE),2)

}
\references{
\itemize{
\item Feizi, S., Marbach, D., Médard, M., & Kellis, M. (2013). Network deconvolution as a general method to distinguish direct dependencies in networks.
    \emph{Nature Biotechnology}, 31(8), 726-733. \url{https://doi.org/10.1038/nbt.2635}
}
}
\author{
\itemize{
 \item Soheil Feizi, KELLIS-LAB, Matlab code
 \item Detlef Groth, University of Potsdam, R code
}
}
