% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.new}
\alias{asg.new}
\title{Create an association chain graph.}
\usage{
asg.new(
  data,
  alpha = 0.05,
  method = "pearson",
  threshold = 0.01,
  pcor = FALSE,
  check.singles = FALSE,
  chains.clean = TRUE,
  prob = FALSE,
  prob.threshold = 0.2,
  prob.n = 25
)
}
\arguments{
\item{data}{A dataframe where network nodes are the rownames and data variables are in the columns.}

\item{alpha}{Confidence threshold for p-value edge cutting after all chains were generated, default: 0.05.}

\item{method}{Method to calculate correlation/association values, can be pearson, spearman, kendall, cor.fk (requires pcaPP package), rpart or mi (mutual information) default: 'pearson'.}

\item{threshold}{R-squared correlation coefficient threshold for which r-square values should be used for chain generation, r=0.1 is r-square of 0.01, default: 0.01.}

\item{pcor}{should edges with high partial correlations added, default: FALSE.}

\item{check.singles}{should isolated nodes connected with sufficent high R^2 and significance, default: FALSE.}

\item{chains.clean}{should shorter chains be removed if they are in longer chains, and should reverse dubplicated chains be removed, default: TRUE}

\item{prob}{should be probabilities computed for each edge using bootstrapping. Only in this case the parameters starting with prob are used, default: FALSE}

\item{prob.threshold}{threshold to set an edge, a value of 0.5 means, that the edge must be found in 50\% of all samplings, default: 0.2}

\item{prob.n}{number of boostrap samples to be taken, default: 25}
}
\value{
An asg graph data object with the fields theta for the adjacency matrix, sigma for the correlation matrix, chains for the association chains and data representing the input data.
}
\description{
This function returns graphs which are build from association chains.
}
\examples{
data(swiss)
sw.g=asg.new(swiss,method='spearman')
# what objects are there?
ls(sw.g)
sw.g$theta
round(sw.g$sigma,2)
par(mfrow=c(1,2))
plot(sw.g,layout='sam')
plot(sw.g,layout='circle')
plot(sw.g,layout='star',star.center='Catholic')
# resampling approach
sw.g=asg.new(swiss,method='spearman',check.singles=TRUE,prob=TRUE)
sw.g$theta
sw.g$probablities
}
\seealso{
\link[asg:plot.asg]{plot.asg}
}
\keyword{correlation}
\keyword{network}
