% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.pcadata}
\alias{asg.pcadata}
\title{Restores data out of a PCA with possible component reduction}
\usage{
asg.pcadata(x, scale = TRUE, reduce = NULL, na.omit = TRUE)
}
\arguments{
\item{x}{A PCA or a data frame or matrix with data.}

\item{scale}{in scale original data are given, should the data being scaled in the PCA}

\item{reduce}{list object with PC1 ... PCn as names and numerical values from 0 and 1 where 1 means keep the data and 0 means the complete removal of the component, default: NULL}

\item{na.omit}{What do with missing values, if TRUE rows  with missing values will be deleted, if FALSE an error is thrown, default: TRUE}
}
\value{
a data matrix with variables in the columns
}
\description{
This function allows the calculation of the original data out of a PCA object with an option to diminish or to even remove the effect of specific components. If an original data frame or matrix is given a PCA is performed and thereafter the data caluclation is done, the latter should be used obviously only if your intention is to remove or reduce the effect of one or more components.
}
\examples{
data(swiss)
head(swiss)
swiss2=asg.pcadata(swiss[,1:4],scale=FALSE)
head(swiss2) # the same
swiss3=asg.pcadata(swiss[,1:4],reduce=list(PC1=0.5))
head(swiss3) # reduced
swiss4=asg.pcadata(swiss[,1:4],reduce=list(PC1=0.0))
head(swiss4) # removed
}
