% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.pcaplot}
\alias{asg.pcaplot}
\title{improved score plot for pca objects}
\usage{
asg.pcaplot(
  pca,
  pcs = c("PC1", "PC2"),
  pch = 19,
  cex = 7,
  col = "salmon",
  text = TRUE,
  xlab = NULL,
  ylab = NULL,
  grid = TRUE,
  scale = NULL,
  ...
)
}
\arguments{
\item{pca}{pca object of class _prcomp_, created using the function _prcomp_.}

\item{pcs}{the components to plot, default: c('PC1','PC2')}

\item{pch}{plotting character, default: 19}

\item{cex}{character expansion, default: 7}

\item{col}{plotting color, default: salmon}

\item{text}{should text of samples be displayed, default: TRUE}

\item{xlab}{custom xlab, if not given the PC name with variance in percent is shown, default: NULL}

\item{ylab}{custom ylab, if not given the PC name with variance in percent is shown, default: NULL}

\item{grid}{should a plot grid be drawn, default: TRUE}

\item{scale}{function to scale to coordinate values sich as asinh, default: NULL}

\item{\ldots}{additional arguments delegated to the standard plot function}
}
\description{
The function `asg.pcaplot` provides an improved xyplot for
  visualizing the 2D scores of individual principal components of 
  an object created using the function _prcomp_.
}
\examples{
set.seed(125)
library(MASS)
data(birthwt)
birthwt$low=NULL # remove column
rnd=round(rnorm(nrow(birthwt),mean=10,sd=2),2)
birthwt=cbind(birthwt,rnd=rnd) # adding a random var
head(birthwt)
pca=prcomp(t(scale(birthwt)))
par(mfrow=c(1,2),mai=c(0.8,0.8,0.4,0.2))
asg.pcaplot(pca,cex=7)
asg.pcaplot(pca,col=rainbow(ncol(birthwt)),cex=1,text=FALSE)
}
\seealso{
\link[asg:plot.asg]{plot.asg}
}
