% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.pcheck}
\alias{asg.pcheck}
\title{Check graph for partial correlations.}
\usage{
asg.pcheck(graph)
}
\arguments{
\item{graph}{an asg graph object}
}
\value{
a adjacency matrix where edges are checked for partial correlations.
}
\description{
This is a utiltiy function to check existing edges in an asg graph object for 
significant partial correlations.
}
\examples{
data(swiss)
sw.g=asg.new(swiss)
M=asg.pcheck(sw.g)
}
