% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.rpart}
\alias{asg.rpart}
\title{Calculate variable importance for decision tree regression}
\usage{
asg.rpart(x, correct = TRUE)
}
\arguments{
\item{x}{a matrix or a dataframe with variables in columns}

\item{correct}{should be bacgkround substraction of a random variable be perfomed, default: TRUE}
}
\value{
matrix with all pairwise variable importances. The response variables are the row names, the predictor variables are in the column names.
}
\description{
This function returns the variable importance for a regression tree predicting each 
variable individually. Can be used to compare correlations with predictions from a regression tree. 
Good predictors with high importance values should be joined by edges within the asg graph.
}
\examples{
data(swiss)
asg.rpart(swiss)
}
