% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.rsquare}
\alias{asg.rsquare}
\title{linear model r-square for given data and graph}
\usage{
asg.rsquare(data, g = NULL)
}
\arguments{
\item{data}{asg graph object or data matrix or data frame where variables are in columns and samples in rows}

\item{g}{graph object or adjacency matrix of an (un)directed graph, not needed if data is an asg graph, default: NULL.}
}
\value{
vector of rsquare values for each node of the graph.
}
\description{
`asg.rsquare` calculates for given data and a graph the covered r-squared by 
a linear model for each node. The linear model predicts each node by an additive mode 
using it's neighbor nodes in the graph.
}
\examples{
# random data
A=matrix(rbinom(100,1, 0.2),nrow=10,ncol=10)
diag(A)=0
colnames(A)=rownames(A)=LETTERS[1:10]
data=matrix(rnorm(1000),ncol=10)
colnames(data)=colnames(A)
asg.rsquare(data,A)
# real data
data(swiss)
sw.s=asg.new(swiss,method='spearman')
rsqs=asg.rsquare(sw.s)
plot(sw.s,main=paste("r =",round(mean(rsqs,2))),
   layout='star',star.center='Examination')
# some colors for r-square values
vcols=paste("grey",seq(80,40,by=-10),sep="")
scols=as.character(cut(asg.rsquare(swiss,sw.s$theta),
   breaks=c(0,0.1,0.3,0.5,0.7,1),labels=vcols))
plot(sw.s,main=paste("r =",round(mean(asg.rsquare(swiss,sw.s$theta)),2)),
   vertex.color=scols ,layout='star',star.center='Examination',
   vertex.size=10,edge.color=c('black','red'),edge.width=3)
}
