% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.shortest.paths}
\alias{asg.shortest.paths}
\title{Get shortest paths between nodes from adjacency matrices or mcgraph objects}
\usage{
asg.shortest.paths(A, mode = "directed")
}
\arguments{
\item{A}{mcgraph object or adjacency matrix}

\item{mode}{should graph be taken as directed graphs, or undirected, if mode='undirected' is given, the graph is transformed into an undirected graph, if mode is 'directed' no graph transformation is done, default: 'directed'}
}
\value{
matrix of same dimension as input matrix with vector with component ids and node names labeling the ids.
}
\description{
Returns the shortes paths between each pair of nodes or Inf if they belong
    to different graph components which are not connected. Attention this only works for undirected graphs and directed graphs are converted internally to undirected graphs. For a full and faster implemention use more sophisticated R packages, like igraph. The method is mainly used for  the layout mechanism.
}
\examples{
data(swiss)
sw.s=asg.new(swiss,method='spearman')
plot(sw.s,layout='sam')
asg.shortest.paths(sw.s)
}
\author{
Detlef Groth <email: dgroth@uni-potsdam.de>
}
