% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{mgraph.autonames}
\alias{mgraph.autonames}
\title{create names for nodes and other data structures}
\usage{
mgraph.autonames(n, prefix = NULL)
}
\arguments{
\item{n}{- how many labels}

\item{prefix}{one or more prefixes to be used, default: NULL.}
}
\description{
This function aids in creating standard node labels for graphs.
}
\examples{
mgraph.autonames(12)
mgraph.autonames(12,LETTERS[1:4])
mgraph.autonames(12,"R")

}
\author{
Masiar Novine, Detlef Groth
}
