% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{mgraph.components}
\alias{mgraph.components}
\title{extract graph components}
\usage{
mgraph.components(g)
}
\arguments{
\item{g}{a mgraph object or an adjacency matrix or an adjacency list}
}
\description{
Return graph components, nodes which are connected are within the same component.
}
\examples{
R = mgraph.new(type="random",nodes=20,edges=25)
mgraph.components(R)
}
