% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{mgraph.d2u}
\alias{mgraph.d2u}
\title{create undirected graph out of a directed graph}
\usage{
mgraph.d2u(g)
}
\arguments{
\item{g}{a mgraph object or an adjacency matrix}
}
\description{
This function gets an directed graph and convertes all edges from directed ones to undirected ones.
  The number of edges should stay the same, the edge sign (+ or -) stays the same.
}
\examples{
par(mfrow=c(1,2),mai=rep(0.2,4))
A=mgraph.new(type="angie",nodes=4,edges=4)
A
U=mgraph.d2u(A)
U
plot(A); plot(U)

}
