% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{mgraph.degree}
\alias{mgraph.degree}
\title{number of undirected or incoming and outgoing edges.}
\usage{
mgraph.degree(g, mode = "undirected")
}
\arguments{
\item{g}{a mgraph object or an adjacency matrix}

\item{mode}{either 'undirected', 'out' or 'in', default: 'undirected'}
}
\description{
This function returns the degree centrality measure for every node.
}
\examples{
par(mfrow=c(1,2),mai=rep(0.2,4))
A=mgraph.new(type="regular",nodes=8,k=3)
plot(A,layout="circle"); 
plot(mgraph.d2u(A),layout="circle"); 
mgraph.degree(A)
mgraph.degree(A,mode="out")
mgraph.degree(A,mode="in")

}
