% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{mgraph.graph2data}
\alias{mgraph.graph2data}
\title{create correlated data for the given graph}
\usage{
mgraph.graph2data(
  g,
  n = 100,
  iter = 15,
  val = 100,
  sd = 2,
  prop = 0.05,
  noise = 1,
  method = "mc"
)
}
\arguments{
\item{g}{a mgraph object or an adjacency matrix}

\item{n}{the number of measurements per node, default: 100}

\item{iter}{the number of iterations, default: 15}

\item{val}{initial node value, default: 100}

\item{sd}{initial standard deviation, default: 2}

\item{prop}{proportion of the target node value take from the source node, default: 0.05}

\item{noise}{the sd for the noise value added after each iteration using rnorm function with mean 0, default: 1}

\item{method}{the method for data generation, either 'mc' for using Monte Carlo simulation or 'pc' for using a precision matrix, default: 'mc'}
}
\description{
This function is a short implementation of the algorithm in Novine et. al. 2022.
}
\examples{
par(mfrow=c(1,2),mai=rep(0.2,4))
A=mgraph.new(type="angie",nodes=8,edges=10)
plot(A,layout="circle"); 
C=mgraph.graph2data(A,prop=0.05)
round(cor(t(C)),2)
# Here an example which shows that with increasing number of
# iterations the correlations between the nodes increase.
par(mfrow=c(2,2),mai=rep(0.1,4))
W=mgraph.new(type="werner")
plot(W,layout='werner')
for (i in c(15,30,50)) {
     data=mgraph.graph2data(W,n=200,iter=i)
     as=asg.new(t(data),method="spearman",alpha=0.1)
     plot(mgraph.new(as$theta),layout='werner',edge.text=round(as$sigma,2))
     text(2.5,1.5,paste("iter =",i),cex=2)
  }

}
