% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{mgraph.harmonic_centrality}
\alias{mgraph.harmonic_centrality}
\title{return the harmonic centraility for each node of a graph}
\usage{
mgraph.harmonic_centrality(g, norm = TRUE)
}
\arguments{
\item{g}{a mgraph object}

\item{norm}{should the harmonic centrality normalized over the number of nodes, default: TRUE}
}
\description{
This function returns the harmonic centrality for every node, of a graph
    directed graphs are converted to unidirecred graphs before and then the following
    formula is applied: \code{h = sum(1/d) / (n-1)} where d are the shortest paths
    of a node to all other nodes of the graph and n ist the number of nodes in a graph.
}
\examples{
B=mgraph.new(type="angie",nodes=20,edges=30)
plot(B,layout="sam")
mgraph.degree(B)
mgraph.harmonic_centrality(B)
cols=mgraph.nodeColors(B,type="harmonic")
plot(B,layout="sam",vertex.color=cols)
lg=mgraph.nodeColors(B,type="harmonic",legend=TRUE)
legend("bottom",fill=lg$col, legend=lg$legend,ncol=6)
}
\references{
\itemize{
\item Marchiori, M., & Latora, V. (2000). Harmony in the small-world. Physica A: Statistical Mechanics and its Applications, 285(3-4), 539-546.
}
}
