% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{mgraph.new}
\alias{mgraph.new}
\title{create mgraph objects}
\usage{
mgraph.new(
  x = NULL,
  type = "random",
  mode = "directed",
  nodes = 10,
  edges = 12,
  m = 1,
  k = 3,
  p = NULL,
  power = 1,
  input = 2
)
}
\arguments{
\item{x}{either a adjacency matrix or an adjacency list, if not given a type must be given}

\item{type}{a graph type, one of 'angie', 'band', 'barabasi', 'circle', 'cluster', 'hubs', 'orion','random', 'regular' or 'werner', default: 'random'}

\item{mode}{either 'undirected' or 'directed', default: 'directed'}

\item{nodes}{number of nodes for a given type, default: 10}

\item{edges}{number of edges for a given type, not used for type "barabasi", default: 12}

\item{m}{number of edges added in each iteration of type is 'barabasi', default: 1}

\item{k}{the degree for a regular graph, or the number of clusters for a cluster graph or the number of hubs for a hubs graph, default: 2}

\item{p}{the probabilty for edges if type is 'gnp'}

\item{power}{the power for preferential attachment if type is 'barabasi', 1 is linear preferential attachment, below one is sublinear, smaller hubs, 0 is no hubs, above 1 is super linear with super hubs, default: 1}

\item{input}{number of input nodes for random graphs, default: 2}
}
\description{
This function creates a new graph object based on either
 an adjacency matrix or for a given graph type.
}
\examples{
M <- matrix(0,nrow=7,ncol=7)
rownames(M)=colnames(M)=LETTERS[1:7]
M[c('A','B'),'C']=1
M['C','D']=1
M['D',c('E','F')]=1
M['E','F']=1
W=mgraph.new(M)
set.seed(125)
R = mgraph.new(type="random",nodes=8,edges=9)
summary(R)
A = mgraph.new(type="angie",nodes=8,edges=9)
C = mgraph.new(type="circle",nodes=8)
par(mfrow=c(3,2),mai=rep(0.3,4))
plot(W,layout="sam",vertex.color=mgraph.nodeColors(W))
plot(R,layout="circle")
plot(A,layout="circle")
plot(C,layout="circle")
O=mgraph.new(type="orion")
plot(O,layout='orion',vertex.colors=mgraph.nodeColors(O))
plot(mgraph.d2u(O),layout="orion")
}
\author{
: Detlef Groth, Cedric Moris
}
