% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{plot.mgraph}
\alias{plot.mgraph}
\title{plot a given graph or adjacency matrix}
\usage{
\method{plot}{mgraph}(x, layout = "sam", ...)
}
\arguments{
\item{x}{a mgraph object or an adjacency matrix or an adjacency list}

\item{layout}{the layout, either a matrix or string in addition to the layout parameters for plot.asg the layout strings 'werner' for 6 node graphs and 'orion' for 20 node graphs are supported, default: 'sam'}

\item{\ldots}{all remaining arguments are forwarded to the asg.mplot function}
}
\description{
plot a given graph or adjacency matrix

This function is a plotting function for mgraph objects. All arguments are just forwarded to the asg.mplot function.
}
\examples{
par(mfrow=c(2,2),mai=rep(0.2,4))
G=mgraph.new(type="angie")
plot(G, layout="sam")
plot(G, layout="circle") 
plot(G, layout="star",vertex.color="grey90") 
M=matrix(0,ncol=5,nrow=5)
diag(M)=0
M[upper.tri(M)]= as.integer(rlnorm(10,0.35))
M[lower.tri(M)]=t(M)[lower.tri(M)]
rownames(M)=colnames(M)=LETTERS[1:5]
M=mgraph.new(M)
M
identical(M,t(M))

}
\seealso{
\link[asg:plot.asg]{plot.asg(x)}
}
\author{
Detlef Groth, Cedric Moris
}
