% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R, R/pkgname.R
\docType{package}
\name{asg-package}
\alias{asg}
\alias{asg-package}
\title{asg package - association chain graphs from correlation networks}
\description{
The asg package can b used to construct association chain graphs based on the St. Nicolas House algorithm as described in Groth et. 2019.

What the package does (one paragraph).
}
\details{
The package provides the following functions:

Function for graph generation from data:
\describe{
\item{\link[asg:asg.new]{asg.new(data)}}{creates and returns a new asg graph}
}

S3 methods for asg graphs:
\describe{
\item{\link[asg:plot.asg]{plot.asg(asggraph)}}{plots an asg graph}
\item{\link[asg:as.list.asg]{as.list.asg(asggraph)}}{return a list representation of an asg graph object}
}

Utility functions:
\describe{
\item{\link[asg:asg.getChains]{asg.getChains(asggraph)}}{returns the chains found by the algorithm as matrix}
\item{\link[asg:asg.getFactors]{asg.getFactors(data,factors=0)}}{performs a factor analysis on the given data}
\item{\link[asg:asg.impute]{asg.impute(x)}}{impute missing values using knn, rpart, mean or median approaches}
\item{\link[asg:asg.layout]{asg.layout(g)}}{calculate layout coordinates for the given graph or adjacency matrix}
\item{\link[asg:asg.ll]{asg.ll(g,chain)}}{calculate log-likelihood for the given chain of the asg graph}
\item{\link[asg:asg.mdsplot]{asg.mdsplot(g)}}{do a MDS plot for the variables of a asg object}
\item{\link[asg:asg.mplot]{asg.mplot(mt)}}{plots an adjacency matrix using the plotting facilities of the asg package}
\item{\link[asg:asg.nd]{asg.nd(x)}}{apply network deconvolution on correlation matrices according to Feizi et al. 2013}
\item{\link[asg:asg.pcaplot]{asg.pcaplot(g)}}{do a PCA score plot for a PCA object}
\item{\link[asg:asg.rsquare]{asg.rsquare(data,g)}}{for given data and graph or adjacency matrix calculate linear model r-square value}
}

More internal and explorative functions. Not required for standard applications:
\describe{
\item{\link[asg:asg.getGraphAccuracy]{asg.getGraphAccuracy(true.graph,pred.graph)}}{how close is the predicted graph to the real graph}
\item{\link[asg:asg.rpart]{asg.rpart(data)}}{variable importance for prediction based on regression trees}
\item{\link[asg:asg.pcadata]{asg.pcadata(x)}}{restore PCA data from a PCA object or reduce some components and then restore the data}
\item{\link[asg:asg.pcheck]{asg.pcheck(asggraph)}}{checks graph for partial correlations between not yet connected nodes which might be added as edges if significant}
}

Functions to create your own graphs:
\describe{
\item{\link[asg:mgraph.new]{mgraph.new(x=NULL,type="angie")}}{function to create different graph types}
\item{\link[asg:plot.mgraph]{plot.mgraph(g)}}{plot function for a S3 mgraph object}
\item{\link[asg:mgraph.autonames]{mgraph.autonames(n,prefix)}}{create names for nodes and other items}
\item{\link[asg:mgraph.components]{mgraph.components(g)}}{extract graph components and the ids for each node}
\item{\link[asg:mgraph.degree]{mgraph.degree(g)}}{get number of edges for a given graph}
\item{\link[asg:mgraph.d2u]{mgraph.d2u(g)}}{convert a directed to an undirected graph}
\item{\link[asg:mgraph.graph2data]{mgraph.graph2data(g)}}{create correlated data for a given graph}
\item{\link[asg:mgraph.nodeColors]{mgraph.nodeColors(g)}}{create color codes based on in- and out-degrees for the given graph}
\item{\link[asg:mgraph.nodeColors]{mgraph.nodeColors(g)}}{create color codes based on in- and out-degrees for the given graph}
\item{\link[asg:mgraph.u2d]{mgraph.u2d(g)}}{convert an undirected to a directed graph}
}

Below is a standard analysis pipeline using the asg graph package. We use here the swiss data as example data:

The only function you're likely to need from asg is [asg()]. 
Otherwise refer to the vignettes to see how to format the documentation.
}
\examples{
data(swiss)
# alpha = 0.1 is recommended for smaller sample sets, default is 0.01
asg=asg.new(swiss,method='spearman',alpha=0.1)
plot(asg,layout='sam')
# we can as well do resamplings
asg=asg.new(swiss,method='spearman',alpha=0.1,prob=TRUE)
plot(asg,layout='sam')
asg$theta
asg$probabilities
}
\references{
\itemize{
   \item Groth, D., Scheffler, C., & Hermanussen, M. (2019). 
   Body height in stunted Indonesian children depends directly on parental education and not 
    via a nutrition mediated pathway - 
    Evidence from tracing association chains by St. Nicolas House Analysis. 
    \emph{Anthropologischer Anzeiger}, 76 No. 5 (2019), p. 445 - 451. \url{https://doi.org/10.1127/anthranz/2019/1027}
   \item Hermanussen, M., Assmann, C. und Groth, D. (2021).
   Chain Reversion for Detecting Associations in Interacting Variables - St. Nicolas House Analysis.
   \emph{International Journal of Environmental Research and Public Health}. 18, 4 (2021). \url{https://doi.org/10.3390/ijerph18041741}.
   \item Novine, M., Mattsson, C. C., & Groth, D. (2021).
   Network reconstruction based on synthetic data generated by a Monte Carlo approach. \emph{Human Biology and Public Health}, 3:26. \url{ https://doi.org/10.52905/hbph2021.3.26}
}
}
\author{
Detlef Groth <dgroth@uni-potsdam.de>

\strong{Maintainer}: Detlef Groth \email{dgroth@uni-potsdam.de} (\href{https://orcid.org/0000-0002-9441-3978}{ORCID})

Authors:
\itemize{
  \item Masiar Novine \email{novinefarahb@uni-potsdam.de} (\href{https://orcid.org/0000-0001-9687-8675}{ORCID})
  \item Cedric Moris \email{cedric.moris@uni-potsdam.de}
}

}
\keyword{internal}
